DNS: avoid setting the state to DONE on a callback error path.
If the state is set to DONE on an error path in dns_recv and the
'found' callback re-enters the DNS code to dns_enqueue then in might
reused this DONE entry but dns_recv sets the state of the entry to
UNUSED after the callback returns, the state is lost and the callback
is never called.

See this happen in the sntp code, it just loses the callback and stops
sending out requests. The upstream code seems to have eliminated the
DONE state so probably does not have this issue, although some uses of
the DONE state remain, see https://savannah.nongnu.org/bugs/index.php?48396